[structure] Refactor and generalize ordering menu item building
